Create valid query for fields sources
